Join league working
